+++ src/main/java/blackjack/domain/player/AbstractPlayer.java
@@ -1,4 +1,4 @@
-package blackjack.domain.gamer;
+package blackjack.domain.player;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
