+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -1,7 +1,6 @@
-package blackjack.domain.gamer;
+package blackjack.domain.player;
 
 import blackjack.domain.WinDrawLose;
-import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import java.util.ArrayList;
 import java.util.EnumMap;
