+++ src/test/java/blackjack/domain/DealerTest.java
@@ -3,8 +3,8 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Deck;
-import blackjack.domain.gamer.Dealer;
-import blackjack.domain.gamer.Player;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
