+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.card.Deck;
-import blackjack.domain.gamer.Player;
-import blackjack.domain.gamer.CasinoCustomer;
+import blackjack.domain.player.Player;
+import blackjack.domain.player.CasinoCustomer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
